Userprofile

+++ dosen/index.php
@@ -1,3 +1,32 @@
+<?php
+session_start();
+$user = $_SESSION['user'];
+$nip = $_SESSION['nip'];
+$nama = $_SESSION['nama'];
+$jurusan = $_SESSION['jurusan'];
+$hakakses = $_SESSION['hakakses'];
+if ($_SESSION['hakakses'] != "dosen") {
+	header("location:../index.php?pesan=noaccess");
+}
+require('../system/dbconn.php');
+require('../system/myfunc.php');
+?>
+
+<?php
+//cek apabila ada data user kurang lengkap lempar ke update profile
+$stmt = $dbsurat->prepare('SELECT * FROM pengguna WHERE nip=?');
+$stmt->bind_param('s', $nip);
+$stmt->execute();
+$result = $stmt->get_result();
+$dhasil = $result->fetch_assoc();
+$nohp = $dhasil['nohp'];
+$email = $dhasil['email'];
+if ($nohp == null or $email == null) {
+	header("location:userprofile-tampil.php");
+}
+
+?>
+
 <!DOCTYPE html>
 <html>
 
@@ -18,208 +47,20 @@
 	<link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700" rel="stylesheet">
 </head>
 
-<!-- location sharing -->
-<?php
-$lokasi = "coming soon ...";
-?>
-
-<!-- akses ke database -->
-<?php require_once('../system/dbconn.php'); ?>
-
-
-<!-- cek session -->
-<?php
-session_start();
-if ($_SESSION['hakakses'] != "dosen") {
-	header("location:../index.php?pesan=noaccess");
-}
-?>
-
-<?php
-$iduser = $_SESSION['iduser'];
-$nip = $_SESSION['nip'];
-$nama = $_SESSION['nama'];
-$status = $_SESSION['status'];
-$jurusan = $_SESSION['jurusan'];
-$role = $_SESSION['role'];
-?>
-
 <body class="hold-transition sidebar-mini">
 	<!-- Site wrapper -->
 	<div class="wrapper">
 		<!-- Navbar -->
-		<nav class="main-header navbar navbar-expand navbar-white navbar-light">
-			<!-- Left navbar links -->
-			<ul class="navbar-nav">
-				<li class="nav-item">
-					<a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"> Menu</i></a>
-				</li>
-			</ul>
-		</nav>
+		<?php
+		require('navbar.php');
+		?>
 		<!-- /.navbar -->
 
 		<!-- Main Sidebar Container -->
-		<aside class="main-sidebar sidebar-dark-primary elevation-4">
-			<!-- Brand Logo -->
-			<a href="" class="brand-link">
-				<img src="../system/uin-malang-logo.png" alt="UIN Malang" class="brand-image img-circle elevation-3" style="opacity: .8">
-				<span class="brand-text font-weight-light">UIN Malang</span>
-			</a>
-
-			<!-- Sidebar -->
-			<div class="sidebar">
-				<!-- Sidebar user (optional)-->
-				<div class="user-panel mt-3 pb-3 mb-3 d-flex">
-					<div class="info">
-						<a href="#" class="d-block"><?php echo $nama; ?></a>
-						<a href="#" class="d-block">NIP : <?php echo $nip; ?></a>
-					</div>
-				</div>
-
-				<!-- Sidebar Menu -->
-				<nav class="mt-2">
-					<ul class="nav nav-pills nav-sidebar flex-column" data-widget="treeview" role="menu" data-accordion="false">
-						<li class="nav-item">
-							<a href="index.php" class="nav-link">
-								<i class="nav-icon fas fa-th"></i>
-								<p>
-									Dashboard
-									<span class="right badge badge-danger"></span>
-								</p>
-							</a>
-						</li>
-						<li class="nav-item">
-							<a href="bimbingan-tampil.php" class="nav-link">
-								<i class="nav-icon fas fa-users"></i>
-								<p>
-									Mhs. Bimbingan
-									<span class="right badge badge-danger"></span>
-								</p>
-							</a>
-						</li>
-						<li class="nav-item">
-							<a href="datasurat-tampil.php" class="nav-link">
-								<i class="nav-icon fas fa-envelope-open"></i>
-								<p>
-									Riwayat Surat
-									<span class="right badge badge-danger"></span>
-								</p>
-							</a>
-						</li>
-						<li class="nav-item">
-							<a href="wfh-isi.php" class="nav-link">
-								<i class="nav-icon fas fa-envelope"></i>
-								<p>
-									Pengajuan WFH
-									<span class="right badge badge-danger"></span>
-								</p>
-							</a>
-						</li>
-						<?php
-						$qoperator = mysqli_query($dbsurat, "SELECT * FROM skpi_operator WHERE kode='$iduser'");
-						$jmldata = mysqli_num_rows($qoperator);
-						if ($jmldata == 1) {
-
-						?>
-							<li class="nav-item has-treeview menu-close">
-								<a href="#" class="nav-link">
-									<i class="nav-icon fas fa-graduation-cap"></i>
-									<p>
-										SKPI
-										<i class="right fas fa-angle-left"></i>
-									</p>
-								</a>
-								<ul class="nav nav-treeview">
-									<li class="nav-item">
-										<a href="skpi-rekap.php" class="nav-link">
-											<i class="nav-icon fas fa-graduation-cap"></i>
-											<p>
-												Rekap Pengajuan SKPI
-												<!--<span class="right badge badge-danger">BARU</span>-->
-											</p>
-										</a>
-									</li>
-								</ul>
-								<ul class="nav nav-treeview">
-									<li class="nav-item">
-										<a href="skpi-isi.php" class="nav-link">
-											<i class="nav-icon fas fa-graduation-cap"></i>
-											<p>
-												Isi data SKPI
-											</p>
-										</a>
-									</li>
-								</ul>
-							</li>
-						<?php
-						}
-						?>
-						<li class="nav-item">
-							<a href="notifikasi-isi.php" class="nav-link">
-								<i class="nav-icon fas fa-bullhorn"></i>
-								<p>
-									Kirim Notifikasi
-									<span class="right badge badge-danger"></span>
-								</p>
-							</a>
-						</li>
-						<li class="nav-item has-treeview menu-close">
-							<a href="#" class="nav-link">
-								<i class="nav-icon fa fa-file"></i>
-								<p>
-									Dokumen
-									<i class="right fas fa-angle-left"></i>
-								</p>
-							</a>
-							<ul class="nav nav-treeview">
-								<li class="nav-item">
-									<a href="http://saintek.uin-malang.ac.id/online/doc/se276-2021.pdf" target="_blank" class="nav-link">
-										<i class="far fa-file-pdf"></i>
-										<p>SE Rektor UIN Malang</p>
-									</a>
-								</li>
-								<li class="nav-item">
-									<a href="http://saintek.uin-malang.ac.id/online/doc/panduandosen.pdf" target="_blank" class="nav-link">
-										<i class="far fa-file-pdf"></i>
-										<p>Panduan Pengajuan WFH</p>
-									</a>
-								</li>
-								<?php if ($iduser == '62007') { ?>
-									<li class="nav-item">
-										<a href="http://saintek.uin-malang.ac.id/online/doc/panduankajur.pdf" target="_blank" class="nav-link">
-											<i class="far fa-file-pdf"></i>
-											<p>Panduan Verifikasi WFH</p>
-										</a>
-									</li>
-								<?php
-								}
-								?>
-							</ul>
-						</li>
-						<li class="nav-item">
-							<a href="https://wa.me/6281234302099" class="nav-link">
-								<i class="nav-icon fas fa-question-circle"></i>
-								<p>
-									Bantuan
-									<span class="right badge badge-danger"></span>
-								</p>
-							</a>
-						</li>
-						<li class="nav-item">
-							<a href="../logout.php" class="nav-link">
-								<i class="nav-icon fas fa-window-close"></i>
-								<p>
-									Keluar
-									<span class="right badge badge-danger"></span>
-								</p>
-							</a>
-						</li>
-					</ul>
-				</nav>
-				<!-- /.sidebar-menu -->
-			</div>
-			<!-- /.sidebar -->
-		</aside>
+		<?php
+		require('sidebar.php');
+		?>
+		<!-- ./Main Sidebar Container -->
 
 		<!-- Content Wrapper. Contains page content -->
 		<div class="content-wrapper">
@@ -228,12 +69,23 @@ $role = $_SESSION['role'];
 				<div class="container-fluid">
 					<div class="row mb-2">
 						<div class="col-sm-6">
-							<h3></h3>
+							<h3>Dashboard</h3>
 						</div>
 					</div>
 				</div><!-- /.container-fluid -->
 			</section>
-
+			<?php
+			if (isset($_GET['pesan'])) {
+				if ($_GET['pesan'] == "success") {
+			?>
+					<div class="alert alert-success alert-dismissible fade show">
+						<button type="button" class="close" data-dismiss="alert">&times;</button>
+						<strong>UPDATE</strong> data berhasil
+					</div>
+			<?php
+				}
+			}
+			?>
 			<!-- pengajuan surat mahasiswa -->
 			<section class="content">
 				<!-- Default box -->
@@ -269,7 +121,7 @@ $role = $_SESSION['role'];
 
 									<!-- PKL Koordinator-->
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM pkl WHERE validatorkoor='$iduser' AND validasikoordinator = 0");
+									$query = mysqli_query($dbsurat, "SELECT * FROM pkl WHERE validatorkoor='$user' AND validasikoordinator = 0");
 									$jmldata = mysqli_num_rows($query);
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['id'];
@@ -307,7 +159,7 @@ $role = $_SESSION['role'];
 
 									<!-- ijin lab -->
 									<?php
-									$query = mysqli_query($dbsurat, "select * from ijinlab where validatordosen='$iduser' AND validasidosen = 0");
+									$query = mysqli_query($dbsurat, "select * from ijinlab where validatordosen='$user' AND validasidosen = 0");
 									$jmldata = mysqli_num_rows($query);
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['no'];
@@ -345,7 +197,7 @@ $role = $_SESSION['role'];
 
 									<!-- ijin penelitian -->
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM ijinpenelitian WHERE validatordosen='$iduser' AND validasidosen = 0");
+									$query = mysqli_query($dbsurat, "SELECT * FROM ijinpenelitian WHERE validatordosen='$user' AND validasidosen = 0");
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['id'];
 										$nim = $data['nim'];
@@ -382,7 +234,7 @@ $role = $_SESSION['role'];
 
 									<!-- peminjaman alat -->
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM peminjamanalat WHERE validatordosen='$iduser' AND validasidosen = 0");
+									$query = mysqli_query($dbsurat, "SELECT * FROM peminjamanalat WHERE validatordosen='$user' AND validasidosen = 0");
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['id'];
 										$nim = $data['nim'];
@@ -419,7 +271,7 @@ $role = $_SESSION['role'];
 
 									<!-- Observasi -->
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM observasi WHERE validatordosen='$iduser' AND validasidosen = 0");
+									$query = mysqli_query($dbsurat, "SELECT * FROM observasi WHERE validatordosen='$user' AND validasidosen = 0");
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['id'];
 										$nim = $data['nim'];
@@ -456,7 +308,7 @@ $role = $_SESSION['role'];
 
 									<!-- pengambilandata -->
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM pengambilandata WHERE validatordosen='$iduser' AND validasidosen = 0");
+									$query = mysqli_query($dbsurat, "SELECT * FROM pengambilandata WHERE validatordosen='$user' AND validasidosen = 0");
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['id'];
 										$nim = $data['nim'];
@@ -493,7 +345,7 @@ $role = $_SESSION['role'];
 
 									<!-- Surat Keterangan -->
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM suket WHERE validatordosen='$iduser' AND validasidosen is null");
+									$query = mysqli_query($dbsurat, "SELECT * FROM suket WHERE validatordosen='$user' AND validasidosen is null");
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['id'];
 										$nim = $data['nim'];
@@ -530,7 +382,7 @@ $role = $_SESSION['role'];
 
 									<!-- SKPI as Dosen PA -->
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM skpi_prestasipenghargaan WHERE verifikator1='$iduser' AND verifikasi1=0 GROUP BY nim");
+									$query = mysqli_query($dbsurat, "SELECT * FROM skpi_prestasipenghargaan WHERE verifikator1='$user' AND verifikasi1=0 GROUP BY nim");
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['no'];
 										$nim = $data['nim'];
@@ -601,12 +453,12 @@ $role = $_SESSION['role'];
 										<th>Akhir WFH</th>
 										<th>Rencana Kerja</th>
 										<th>Surat Tugas</th>
-										<th>Hapus</th>
+										<th>Aksi</th>
 									</tr>
 								</thead>
 								<tbody>
 									<?php
-									$query = mysqli_query($dbsurat, "SELECT * FROM wfh WHERE iduser='$iduser' ORDER BY tglwfh1 DESC");
+									$query = mysqli_query($dbsurat, "SELECT * FROM wfh WHERE iduser='$user' ORDER BY tglwfh1 DESC");
 									while ($data = mysqli_fetch_array($query)) {
 										$nodata = $data['no'];
 										$tglwfh1 = $data['tglwfh1'];
@@ -713,10 +565,16 @@ $role = $_SESSION['role'];
 												?>
 											</td>
 											<td>
-												<a class="btn btn-danger btn-sm" onclick="return confirm('Yakin menghapus pengajuan ini ?')" href="wfh-hapus.php?nodata=<?php echo $nodata; ?>">
-													<i class="fas fa-trash">
-													</i>
-												</a>
+												<?php
+												if ($verifikasijurusan == 2 or $verifikasifakultas == 2) {
+												?>
+													<a class="btn btn-danger btn-sm" onclick="return confirm('Yakin menghapus pengajuan ini ?')" href="wfh-hapus.php?nodata=<?php echo $nodata; ?>">
+														<i class="fas fa-trash">
+														</i>
+													</a>
+												<?php
+												}
+												?>
 											</td>
 										</tr>
 									<?php
@@ -736,7 +594,9 @@ $role = $_SESSION['role'];
 		<!-- /.content-wrapper -->
 
 		<!-- footer -->
-		<?php include '../system/footerdsn.html' ?>
+		<?php
+		include('footerdsn.html');
+		?>
 		<!-- /.footer -->
 
 		<!-- Control Sidebar -->
